Fixing issue where 'click to edit' message appeared if the user doesn't have permission.
When image has no byline, if the user doesn't have edit permission, then no html is rendered. If the user does have edit permission, then the 'click to edit' message is rendered and the icon appears on hover.

When image has a byline, if the user doesn't have edit permission, then the byline is rendered. If the user does have edit permissions, then the edit icon appears on hover.