Fixing issue where a missing field is not treated correctly.
In a selection of images, its not guaranteed that all images have the same metadata fields, e.g. some may not have a title. `allFields` is the set of all unique metadata fields in the selection of images. Now, we loop over `allFields` to find the corresponding value or undefined for an image.

For example, if there are two images selected and one has a title and the other doesn't, the `metadata` object now looks like `{"title": Set {"title 1", undefined}` rather then `{"title: Set {"title 1"}`. As we switch on the size of `metadata`, we now correctly display "Multiple Titles" rather than "title 1".