symlink config generators to `scripts` directory at project root.
The generators are currently in the `docker` directory, which isn't the most obvious place if you're running the Grid w/out Docker.

Symlinked because there are generators that are specific to Docker (imgops, nginx, osx_hosts_file) within the generators package, so wouldn't make sense to move the generators directory to project root.