Use the Apache Commons HttpClient lib in Cropper (#1926)
... to talk to media api.

There is a bug in Play WS that prevents requests made to secure
endpoints using SNI from handshaking. Cropper is now behind an SNI
secured endpoint and is conseqnetly suffering from this bug:

https://groups.google.com/forum/#!topic/play-framework/T7ZhclgAAMU

This PR moves the request functionality in Cropper to use Apache Commons
HttpClient library which supports SNI as of version 4.3.2, see:

https://issues.apache.org/jira/browse/HTTPCLIENT-1119

This PR is required to move to using Cloudfront for all user facing REST
APIs (in order that out-of-region request times are sped up, and network
reliability increased).