Fix optional field deserialisation bug for userMetadata
This seems to have been caused by a change in behaviour of the `.asOpt` method in Play. Before the upgrade to 2.6 if the field was missing it deserialised to a fully empty `Edits` object, presumably because everything has a default parameter including `ImageMetadata`. After the upgrade it was correctly returning `None`. Unfortunately the frontend now relies on that field being present (and empty if no userMetadata is present in Elasticsearch).