persistable collections
Add configuration for a list of persistable collections. Images in one of these collections will be retained forever.

The primary use-case is for Archive. Currently, they're adding images to the `GNM Archive` collection and adding the `gnmarchive` label and finally manually archive the images. A lot of work! Automating the persistence feels like a quick win.