remove syndicated usages from the syndication tier
When we send an image for syndication, we record a syndication usage.

In order to not send the same image again and again, remove images with a syndication usage from the syndication tier.

This is a slightly hacky solution as it doesn't scale beyond one syndication partner. For example, if partner A takes an image, partner B will no longer see it. However, we only have 1 partner at the moment.

A scalable solution would be to record a syndication key against an usage and filter accordingly. We could do this once we have to support multiple syndication partners.