log endpoint to measure usage
The imports endpoint in image-loader currently accepts any url from the internet, which is only a little scary! This endpoint should only be used by S3 Watcher, but log out who is using it to be sure.

Ultimately, this endpoint will be rewritten to accept an S3 object path rather than a URL. This:
- will reduce the amount of logic in S3 Watcher
- be more secure as only images from the S3 Watcher inject bucket will be accepted
- enable reindexing to reuse this route