allow credit to have no slash
Update BylineCreditReorganise cleaner, allowing credit to be a single element.

This allows the byline to be `foo/bar` and credit to be `foo`. This was previously failing as `credit` had no tail.

Ideally this cleaner would produce `byline:None` and `credit:foo/bar` however it would require a bit of a refactor to support an `Option[String]` type...maybe later?