WIP Test for supplier usages end point.
Addition of new test images regresses some of the existing syndication tests.
Removed mock of DateTime.now as the significance of the 2018-03 is not made explicit (probably the blocking feature?);
better from a documentation point of view to alter the fixtures to allow tests set meaningful dates of test data.