allow cropType to be cleared
Setting the `cropType` query string will set a value in session storage, which is later read to freeze the crop type available. This is useful in embedded Grid when the ratio matters, for example a video thumbnail should be a 16:9 ratio.

Allow session storage to be cleared if `cropType === 'all'`. That is, we're happy to crop to any ratio.