fix odd date issue
When performing ES date aggregations, we're using the month as the interval.

Our test images would sometimes span two months and sometimes one month, this is because we were doing:
- Image A uploaded today
- Image B uploaded 25 days ago

Between 26th and 31st of every month, this test would fail because the number of monthly buckets would be one. However from 1st to 25th of every month, there would be two monthly buckets and the tests would pass.

Image B was a fairly recent addition, hence we've never seen this behaviour until now.