invert over-quota to under-quota
It is more common to ask "what images can I use (under-quota)" rather than "what images can I not use (over-quota)".

Whilst the former could be expressed as not(over-quota), it is by far the more common query so should be the default; to get over-quota we now do not(under-quota).