update date parsing
`.getParser` does not appear to obey the timezone, that is BST was getting parsed into the local timezone. This is incorrect, so add a timezone offset for BST (we're assuming British Summer Time); this means we have to attempt parsing ourselves.

Also stores dates as ISODateTimeFormat because its easier to reason about.