Fail fast when image not found.
Previously, we were always trying to parse the response from elasticsearch into an Image and logging at error level when it failed.

This meant that 404s produce noisy logs relating to json parsing failure which isn't helpful.

Now, we only try to parse the response to an Image if the request succedded (that is, we recieved a hit from elasticsearch).