prevent image download with right click
By adding `pointer-events: none` to the img element, the browser pushes pointer events to the element below the img. That is, when you right click on the image, you're no longer able to download it.

As a result, we need to push the drag-data to a parent element so that you can still drag the image into Composer, InDesign etc.