add a cheat to the bulk insert test
Ideally, before we bulk insert we would assert document count is 0 and after, it is 2. However there is a race condition in the tests (only seen in GH Actions) which results in state being shared between tests, so we're just checking the document count has increased by the correct amount.