ensure we delete files through all code paths in image-loader
There are 3 ways to interact with image loader:
- POST a file, which will create a temp file from the request body
- POST a url, which will download the file from the url into a temp file
- GET a projection, which will create temp files from a file in S3

Each of these paths need to clean up after themselves. We were not doing this on the projection path.

`fromUploadRequestShared` is shared between all three paths, so it makes sense to do the deletion there to keep things DRY.

This has been tested by:
- watching the temp directory (`watch -n 0.1 /tmp`)
- uploading a file through the browser
- observing temp files being created and removed
- projecting an image
- observing temp files being created and removed

Prior to this change, the last step was failing.