change signature of loadImage to use stricter types
Rather than accepting an `Option[String]` as a param and then converting it to a DateTime in the function body, have a stricter signature.

This is part of a refactor to make the ImageLoaderController easier to follow. The ambition is to make the controller a transforming layer, going from an http request into a strictly typed set of objects that can be pushed into other functions.