ship logs to a local ELK stack
Ship logs over TCP to a Logstash instance running locally with https://github.com/guardian/local-elk. This makes the observation of markers a lot easier and provides a smaller feedback loop as we don't need to deploy to TEST to confirm a marker is being written correctly.

Adds a new `--ship-logs` flag to `dev-start.sh` to conditionally ship logs. By default, logs are not shipped. Logs are only shipped when `stage=DEV` and the `--ship-logs` flag is provided.

Logs continue to get written to disk too.