mime type fallback
For some files, Tika detects the mime type as `message/rfc822`. This change falls back to metadata extractor when this happens.

This code path is executed on file upload (or projection) and we've somehow landed in a situation where we have 52 images in PROD with mime type `message/rfc822`. I think the flow was:
- In 2015, upload images
- In 2015, mime type was detected as `image/jpeg`
- In 2019, update mime type detection to use tika only (tika recognises these images as `message/rfc822`)
- In 2020, reingest the images as part of the great image restoration and mime type updated to `message/rfc822`

We should be able to convert these 52 images back to `image/jpeg` by reingesting them.