don't throw when encountering an unrecognised mime type
We've managed to create a number of documents in ES with mime type `message/rfc822`. We use the mime type to create the filename for the content disposition property of the file download link.

Currently, we're throwing (and causing a 5XX response) when we encounter an unsupported mime type. This doesn't make sense at this point - we've already ingested the image, so there's no reason to break downloading the file again.

This changes the behaviour: rather than throwing, we return an empty string which results in the downloaded file having no file extension.