set AWS_CBOR_DISABLE in build.sbt
We need to disable CBOR when using the Java SDK to interact with kinesis running in localstack.

Setting this env var in build.sbt makes it easier to run tests in your IDE as you don't have to remember to set it each time. Also allows you to run `sbt test` without needing to remember to set the slightly magical env var too.