use exact match for nested queries
Previously we were using a `MatchValue` which would result in the use of the englishSStemmerAnalyzer analyzer. This analyzer does a number of things, one of which is ignoring forward slashes ("/"). It's a fuzzy match.

When searching for a usage reference (against field `usages.references.uri` or `usages.references.name`), we'd want to be able to search by Composer URL. The `MatchValue` doesn't work here.

`ExactMatchValue` is actually preferred in a nested query; you'd use a nested query when looking for a specific value, thus performing a fuzzy match is less performant than a term match.