Use github contribution features more effectively
With increasing numbers of contributors it is becoming important
to have a clear record of our submission expectations and make it
easier to do the right thing.

This commit adds an initial set of contribution guidelines and
also tweaks the issue and PR templates. Finally we add a code
owners file to help ensure that when changes are made they are
reviewed by the right people.