Refactor ESinfo to handle simultaneous successes and failures
We need this as the previous assumption - that an image can only succeed
or fail - hasn't held up in reality. If we abandon a migration, an image
that succeeded in that migration may fail in the next.  If this happens,
then query for migration failures will find the image, but the failure
reason won't be returned when the document is parsed. Instead of
encoding the migration field as an Either, use a more standard case
class with both fields available. Plus, this means we get to remove all
of the custom parsing logic required by treating the field as an Either.