Delete thumbnails after upload
Most generated assets are placed into a temp dir that is deleted
after the request is finished, with the exception of the thumbnails.
This means that thumbnails will stick around on disk after request
is finished. Normally this isn't a problem, as thumbnails are small
but during a migration they will add up and can even fill up the disk,
causing more problems.

Move the generated thumbnails inside this temp dir that is scoped to the
lifetime of the request, so that they will also be cleared on completion