Do not log to stdout when running locally
There is a high volume of output when running the grid locally (due to
running multiple microservices in the same session), which are difficult
to grok. In addition (anecdotally) running with stdout logging on
reduces the ability to interact with the sbt shell. I found it very
difficult to get sbt to respond to my ctrl-C inputs, which can be a
symptom of the terminal emulator struggling to keep up with the volume
of output.

This PR disables stdout logging when running locally (ie. when stage ==
DEV), which I have found (anecdotally) to significantly improve
responsiveness from sbt.