Support removing all items from a list metadata field
Values from image metadata can come from two sources: metadata on the
original uploaded image, or user edits via Grid UI. These two sources
are stored as separate fields in our Elasticsearch index, and combined
by Grid application code. When they are merged, user edits are preferred
and should be displayed over original metadata.

In our Scala model, we model most of these fields as Options which makes
sense: either the field has not been set, or it has (including user
edits explicitly setting a field to "", which Grid understands as
"explicitly unset"). However, the fields which can contain multiple
values are modeled as Lists (and 1 Set). This can model a field that has
not been set, or has been set with values, but _cannot model a field
that has been explicitly set to no values_. For example, this means that
a user cannot remove all people from an image where the people have been
identified from the original metadata --- attempting to do so will be
treated as if the user has set no override and the original people will
be shown again.

This PR instead models these list fields as Option[List] (and
Option[Set]), which now allows us to model the difference, as None
represents unset, and Some(Nil) represents explicitly set to empty.