Inline Box library
The frontend repo is updating to scala 2.13. As part of the update we found that Box is not compatible with 2.13 and that the lib also contained a bug. The advice we got was to deprecate the Box lib and inline the Box class in projects that are using it. In preparation for MAPI updating to scala 2.13, this commit removes Box as a dependency and inlines the functionality that MAPI requires.