Create request ID in the RequestLoggingFilter
Currently request IDs are (sometimes) created in the controllers and (sometimes) passed down the execution flow as log markers from there. But this means that the request ID is not included on the access logs, so it becomes _very_ hard to correlate the request in the access logs with other log messages.

This commit changes the Play Filter responsible for the access logs to also create the request ID, and pass it through to the rest of the application via a custom internal-only header `x-grid-request-id`. This new header is only consumed in image-loader so far as a proof of concept, it should be fairly easy to roll out further if this works well!

This commit also removes the UUID from the UploadRequest case class as it wasn't being used