Fix bootstrapping of nonFree query param
The only place in kahuna which should pay attention to the showPaid
permission should be the state for the root (ie. '/'), which after
finding the permission and setting the nonFree state param will
insert that to session storage ang then transition to /search.
If we are on the /search state and do not have nonFree in session
storage, then the user has followed a link directly to a search - in
that case we should respect the current state of the URL (ie.
stateParams.nonFree)