asynchronously run projection for manual migration
Currently, projections for images are run in the migration stream before
the manual migration source is merged in. This has the effect that
projections must be run before the handler for the manual migration
submission returns - if this is a large image then users clicking the
button will be left waiting for several seconds before receiving a
completion message. Instead, merge the streams at the point where we
have an image ID and a ES document version, and _then_ run the migration
- users get much faster feedback. It also makes setting up an endpoint
to run lots of manual migrations (ie. batch retry of failures) much
easier to set up.