Add button to retry all failed migrations on page
Currently the thrall dashboard will show error details for all images
that have failed to migrate, along with a button for each one. However
even intermittent failures may catch several hundred images in one go -
asking an admin to click a button for every single one is not realistic.
In the past we've gotten around this by doing mass edits of ES documents
to make the images eligible for retry, but this isn't great SOP either.

Add a new endpoint and button that queues a page of failures for retry,
which should massively reduce the burden on admins' wrists.