Load crops data from media-api rather than cropper
All other image data is loaded from media-api (ie. datasource is ES), except
for crops, which is fetched from cropper (ie. datasource is S3). This can cause oddities,
for example if a cropper operation fails part-way through; some crop widths
can be exported to S3 but not written to ES, which causes problems when trying
to eg. download from media-api.

I couldn't see a strong reason to have crops behave in this unique fashion, so
rewrote to fetch the data from media-api instead. The only difference between the
two was that the crop download urls were only added to the cropper response; add them
also into the image transformers in media-api