IMAGEDAM-1400 - Fixing 'Social Media' and 'R&M Photographer - staff' templates
Both the Social Media and R&M Photographer - staff templates were not working as they could not be applied.

This was due to the leases for these templates being undefined as they did not have leases associated with them.

The solution was to add an if-statement on line 153 where the undefined error was occurring, that way a lease is not attempted to be accessed if the template does not contain any.