IMAGEDAM-1593 - Updating ElasticSearch with new usage status WIP
Currently, the usage status update successfully updates dynamoDB but not ElasticSearch.

In order for ElasticSearch to be updated, the correct message, containing an updated UsageNotice must be published to Thrall, via kinesis. Thrall will then process this message and run a script that will update ElasticSearch.

The 'plumbing' for flow has been added, all that is left is to implement the functionality to update ElasticSearch