IMAGEDAM-1593 - Updating ElasticSearch with new usage status WIP pt2
Further additions to the ElasticSearch file, specifically the UpdateSingleUsageMessage function.

Attempted to pass usageId and usageParameters in as two tuples as 'prepareUpdateRequest' suggests that the 'params' input can have variable arguments, but was getting a syntax error. Instead choosing to just pass in usageParameters and accessing the usageId from there (if this works can remove usageId from previous functions).

Some modification to the script still required I think, tested with current changes and usage status is still not updating.