IMAGEDAM-1593 - Updating ElasticSearch with new usage status WIP pt2.4
Completed the work that allows for an individual usage status to be updated. The issue with Options being passed into the painless script turned out to not be the source of my problems, I was actually accessing the parameters incorrectly, with this resolved the script worked without any issues.

Also cleaned up the code changes, renaming functions and parameters, so they are accurate and removing various print statements added for debugging.