IMAGEDAM-1593 - Amending updateUsageStatus endpoint to be asynchronous
Use of 'queryByUsage' meant that a Future[Status] was returned by this endpoint instead of the 'Result' type, required 'Authentication.request'. A placeholder return status was originally included so that this wouldn't block development, but once the desired functionality was achieved a permanent solution was desired.

This solution was to make the endpoint explicitly asynchronous, by using 'async.auth', by default this corrected the success case and the failure case was solved by wrapping the output in a Future[]