IMAGEDAM-1502: Media API changes WIP
Made postToUsages more generic, so it can handle a syndicated invocation as well as download. This means that postToCaptureUsage can be removed and setCaptureUsage can now be used for any type of syndication request, and has been renamed to account for this. Some work left to do to clean up 'syndicateImage' as it has comments left over from Ebu's spike.