IMAGEDAM-1502: Media API & JavaScript changes WIP
Making the required changes such that a syndication usage can be set on an image from the 'send to capture button'. Setting this usage will trigger the outbound capture step function and send the image and sidecar to Capture, allowing it to be on the photo sales system.

Still need to test with multiple images and on the INT environment.