IMAEGEDAM-1667: PhotoSales usage appears in usages panel
Previously the photo sales usage was presented in the usages panel based on the existing logic for a syndication usage. Now a custom photo sales usage is rendered instead. React has been used to present this usage and can easily be extended to other new or existing usages if we wish to convert more of the application to React.