IMAEGEDAM-1675: added to photo sales option on usages@platform chip
In order to filter for images that have been sent to photo sales, a new option had to be added to the 'usages@platform' chip - 'Added to Photo Sales'.

On the backend, this query had to be mapped to the 'syndication' platform, so that the filter could actually be carried out. A match is used in anticipation of a 'Removed from Photo Sales' option in the future.