Fix test running in IntelliJ UI by resolving ScalaTest 3.1 vs 3.2 conflict.
Downgrade "org.scalatestplus" %% "mockito-3-4" to 3.1 series to match the 3.1 ScalaTest pulled in by scalatestplus-play.

Resolves a scalac -deprecation snag which breaks running of tests in IntelliJ.
Warns about a deprecation on .right.value but then doesn't know about the new method because ScalaTest 3.2 is not available on the IntelliJ classpath.

This commit works by pushing the transitive ScalaTest from "org.scalatestplus" %% "mockito-3-4" back down to 3.1 for compatibility with "scalatestplus-play".

Notably "scalatestplus-play" cannot be leveled to 6.0.0 / ScalaTest 3.2 without moving to Scala 2.13 because it has not been released for 2.12.