Remove outdated todo comment
This comment [was added in 2018](https://github.com/guardian/grid/commit/9bc68100e26a1978f9e0b4eeda9cb36bbda972fd), when the list of library dependencies
looked pretty different. Generally we'd assume that the comment refers
to the next line after the comment, but this line was a completely
different import in 2018.

Keeping the inline comment seems unhelpful when it no longer points to
an actionable issue. An alternative would be to update the comment so
that it's clearer (or address the issue it signalled) but given that
it's been 6 years it's not obvious that the work to do this would be
worthwhile.