is softDeleted clause might be able to be run in 1/3 the time.
When examined in the profile output, the exists `softDeletedMetadata` clause seems to get expanded into this:
`"description" : "-ConstantScore(DocValuesFieldExistsQuery [field=softDeletedMetadata.deletedBy] DocValuesFieldExistsQuery [field=softDeletedMetadata.deleteTime]) #*:*",`

This probably means if any of the indexed fields of `softDeletedMetadata` exist then so does `softDeletedMetadata`.
If we assume that `softDeletedMetadata.deletedTime` exists for all softDeletedMetadata records then the profiled `time_in_nanos` drops to 1/3 of the original value.

Hard to show real world results in a dev machine where the index fits in a single node's RAM, but in theory this might speed up the expensive count all and find first page of all queries on the initial page load.

Given that the initial page load time is single digit seconds, this might have a real user facing impact.

Specifying a specific field name is more fragile than the current clause to the improvement needs to be worth it.