shutdown metrics actors more gracefully
On shutdown, Play sends the metrics actors a shutdown message, but as it uses an 'ask', it waits for a response. When no response arrives, it assumes something has gone wrong and shuts down with a more severe error.

In fact nothing has gone wrong (usually), we simply don't bother to send a response. Let's do that, both to be graceful, and to enable slightly faster shutdowns